Backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
Flask-JWT-Extended
Werkzeug
openai
python-dotenv
openai-whisper
